requirements.txt
uvicorn
langgraph  
langchain-core  
langchain-community
pydantic
python-dotenv
langhchain_google_genai
gunicorn
flask